Автопостинг вынесен в отдельный файл. Добавлен файл helpers.py.
