+++ UniversityScheduleBot.py
@@ -1,14 +1,13 @@
 # -*- coding: utf-8 -*-
 import logging
 import re
-import threading
 from datetime import datetime, time, timedelta
-from time import sleep
 
 import telebot
-from telebot import types, apihelper
+from telebot import types
 
-from config import config, daysOfWeek, ScheduleType
+from config import config
+from helpers import daysOfWeek, ScheduleType, get_date_keyboard
 from scheduleCreator import create_schedule_text
 from scheduledb import ScheduleDB, organization_field_length, faculty_field_length
 
@@ -52,19 +51,6 @@ commands = {  # Описание команд используещееся в к
     'auto_posting_on <ЧЧ:ММ>': 'Включение и выбор времени для автоматической отправки расписания в диалог',
     'auto_posting_off': 'Выключение автоматической отправки расписания'
 }
-
-
-def get_date_keyboard():
-    date_select = types.ReplyKeyboardMarkup(row_width=2, resize_keyboard=True, one_time_keyboard=False)
-
-    date_select.row("Сегодня")
-    date_select.row("Завтра")
-    date_select.row("Вся неделя")
-    date_select.row("Понедельник", "Вторник")
-    date_select.row("Среда", "Четверг")
-    date_select.row("Пятница", "Суббота")
-
-    return date_select
 
 
 # handle the "/registration" command
@@ -429,80 +415,6 @@ def response_msg(m):
             logger.info('unknown message: {0}'.format(m.text), extra={'userid': m.chat.id})
 
         bot.send_message(cid, "Неизвестная команда", reply_markup=get_date_keyboard())
-
-
-def auto_posting(current_time):
-    today = datetime.now()
-    week_type = (today.isocalendar()[1] + int(config["WEEK_TYPE"])) % 2
-
-    if datetime.weekday(today) == 6:
-        today += timedelta(days=1)
-        week_type = (week_type + 1) % 2
-
-    day = [daysOfWeek[datetime.weekday(today)]]
-
-    # Выборка пользователей из базы у которых установлена отправка расписния на текущий день
-    with ScheduleDB(config) as db:
-        users = db.find_users_where(auto_posting_time=current_time, is_today=True)
-
-    if users is None:
-        return None
-    try:
-        for user in users:
-            cid = user[0]
-            tag = user[1]
-
-            schedule = create_schedule_text(tag, day[0], week_type)
-            if len(schedule[0]) <= 14:
-                continue
-            bot.send_message(cid, schedule, reply_markup=get_date_keyboard())
-
-            # Статистика
-            if config['STATISTIC_TOKEN'] != '':
-                track(config['STATISTIC_TOKEN'], cid, current_time, 'auto_posting')
-            else:
-                logger.info('auto_posting. Time: {0}'.format(current_time), extra={'userid': cid})
-    except BaseException as e:
-        logger.warning('auto_posting: {0}'.format(str(e)), extra={'userid': 0})
-
-    # Выборка пользователей из базы у которых установлена отправка расписния на завтрашний день, 
-    # если сегодня воскресенье, то расписание будет отправляться на понедельник.
-    if datetime.weekday(datetime.now()) != 6:
-        today += timedelta(days=1)
-      
-    day = [daysOfWeek[datetime.weekday(today)]]
-
-    with ScheduleDB(config) as db:
-        users = db.find_users_where(auto_posting_time=current_time, is_today=False)
-
-    if users is None:
-        return None
-    try:
-        for user in users:
-            cid = user[0]
-            tag = user[1]
-
-            schedule = create_schedule_text(tag, day[0], week_type)
-            if len(schedule[0]) <= 14:
-                continue
-            bot.send_message(cid, schedule, reply_markup=get_date_keyboard())
-
-            # Статистика
-            if config['STATISTIC_TOKEN'] != '':
-                track(config['STATISTIC_TOKEN'], cid, current_time, 'auto_posting')
-            else:
-                logger.info('auto_posting. Time: {0}'.format(current_time), extra={'userid': cid})
-    except BaseException as e:
-        logger.warning('auto_posting: {0}'.format(str(e)), extra={'userid': 0})
-
-
-def auto_posting_thread():
-    while True:
-        threading.Thread(target=auto_posting(datetime.now().time().strftime("%H:%M:00"))).start()
-        # Вычисляем разницу в секундах, между началом минуты и временем завершения потока
-        time_delta = datetime.now() - datetime.now().replace(second=0, microsecond=0)
-        # Поток засыпает на время равное количеству секунд до следующей минуты
-        sleep(60 - time_delta.seconds)
 
 
 # Remove webhook, it fails sometimes the set if there is a previous webhook
